requirements.txt
fastapi==0.115.12
openai==1.77.0
pydantic==2.11.4
pydantic-settings==2.9.1
python-dotenv==1.1.0
uvicorn==0.34.2
mysql